To speed up the efficiency of Mat conversion to go image, move ToImage related functions from core.go to imgporc.go, because CvtColor is needed.
