+++ core.go
@@ -745,106 +745,6 @@ func (m *Mat) T() Mat {
 	return newMat(C.Mat_T(m.p))
 }
 
-// ToImage converts a Mat to a image.Image.
-func (m *Mat) ToImage() (image.Image, error) {
-	t := m.Type()
-	if t != MatTypeCV8UC1 && t != MatTypeCV8UC3 && t != MatTypeCV8UC4 {
-		return nil, errors.New("ToImage supports only MatType CV8UC1, CV8UC3 and CV8UC4")
-	}
-
-	width := m.Cols()
-	height := m.Rows()
-	step := m.Step()
-	data := m.ToBytes()
-	channels := m.Channels()
-
-	if t == MatTypeCV8UC1 {
-		img := image.NewGray(image.Rect(0, 0, width, height))
-		c := color.Gray{Y: uint8(0)}
-
-		for y := 0; y < height; y++ {
-			for x := 0; x < width; x++ {
-				c.Y = uint8(data[y*step+x])
-				img.SetGray(x, y, c)
-			}
-		}
-
-		return img, nil
-	}
-
-	img := image.NewRGBA(image.Rect(0, 0, width, height))
-	c := color.RGBA{
-		R: uint8(0),
-		G: uint8(0),
-		B: uint8(0),
-		A: uint8(255),
-	}
-
-	for y := 0; y < height; y++ {
-		for x := 0; x < step; x = x + channels {
-			c.B = uint8(data[y*step+x])
-			c.G = uint8(data[y*step+x+1])
-			c.R = uint8(data[y*step+x+2])
-			if channels == 4 {
-				c.A = uint8(data[y*step+x+3])
-			}
-			img.SetRGBA(int(x/channels), y, c)
-		}
-	}
-
-	return img, nil
-}
-
-//ImageToMatRGBA converts image.Image to gocv.Mat,
-//which represents RGBA image having 8bit for each component.
-//Type of Mat is gocv.MatTypeCV8UC4.
-func ImageToMatRGBA(img image.Image) (Mat, error) {
-	bounds := img.Bounds()
-	x := bounds.Dx()
-	y := bounds.Dy()
-	data := make([]byte, 0, x*y*4)
-	for j := bounds.Min.Y; j < bounds.Max.Y; j++ {
-		for i := bounds.Min.X; i < bounds.Max.X; i++ {
-			r, g, b, a := img.At(i, j).RGBA()
-			data = append(data, byte(b>>8), byte(g>>8), byte(r>>8), byte(a>>8))
-		}
-	}
-	return NewMatFromBytes(y, x, MatTypeCV8UC4, data)
-}
-
-//ImageToMatRGB converts image.Image to gocv.Mat,
-//which represents RGB image having 8bit for each component.
-//Type of Mat is gocv.MatTypeCV8UC3.
-func ImageToMatRGB(img image.Image) (Mat, error) {
-	bounds := img.Bounds()
-	x := bounds.Dx()
-	y := bounds.Dy()
-	data := make([]byte, 0, x*y*3)
-	for j := bounds.Min.Y; j < bounds.Max.Y; j++ {
-		for i := bounds.Min.X; i < bounds.Max.X; i++ {
-			r, g, b, _ := img.At(i, j).RGBA()
-			data = append(data, byte(b>>8), byte(g>>8), byte(r>>8))
-		}
-	}
-	return NewMatFromBytes(y, x, MatTypeCV8UC3, data)
-}
-
-//ImageGrayToMatGray converts image.Gray to gocv.Mat,
-//which represents grayscale image 8bit.
-//Type of Mat is gocv.MatTypeCV8UC1.
-func ImageGrayToMatGray(img *image.Gray) (Mat, error) {
-	bounds := img.Bounds()
-	x := bounds.Dx()
-	y := bounds.Dy()
-	data := make([]byte, 0, x*y)
-	for j := bounds.Min.Y; j < bounds.Max.Y; j++ {
-		for i := bounds.Min.X; i < bounds.Max.X; i++ {
-			data = append(data, img.GrayAt(i, j).Y)
-		}
-	}
-	return NewMatFromBytes(y, x, MatTypeCV8UC1, data)
-}
-
 // AbsDiff calculates the per-element absolute difference between two arrays
 // or between an array and a scalar.
 //
